Load LLaVA-NeXT in bfloat16 instead of float16 to fix greedy-decoding EOS-collapse from LLaMA-3 fp16 numerical instability

+++ llava_next_visual_attack.py
@@ -150,7 +150,7 @@ def main():
         return
 
     model = LlavaNextForConditionalGeneration.from_pretrained(
-        args.model_id, torch_dtype=torch.float16, device_map=None,
+        args.model_id, torch_dtype=torch.bfloat16, device_map=None,
     ).to(device)
     model.requires_grad_(False)
     # Gradient checkpointing trades compute for activation memory (recomputes
